WINDUP-1231 - Moved persisted model from user's workspace to model plugin's installation location. This will then get deleted during uninstallation.

+++ plugins/org.jboss.tools.windup.ui/src/org/jboss/tools/windup/ui/internal/services/LaunchConfigurationService.java
@@ -60,10 +60,12 @@ public class LaunchConfigurationService implements ILaunchConfigurationListener 
 	public void launchConfigurationRemoved(ILaunchConfiguration launchConfig) {
 		ConfigurationElement configuration = modelService.findConfiguration(launchConfig.getName());
 		ConfigurationElement lastConfiguration = modelService.getRecentConfiguration();
-		if (configuration.getName().equals(lastConfiguration.getName())) {
-			markerService.deleteAllWindupMarkers();
+		if (configuration != null) {
+			if (configuration.getName().equals(lastConfiguration.getName())) {
+				markerService.deleteAllWindupMarkers();
+			}
+			modelService.deleteConfiguration(configuration);
 		}
-		modelService.deleteConfiguration(configuration);
 	}
 	
 	private boolean isWindupConfig(ILaunchConfiguration launchConfig) {
